+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## X.X.X (YYYY-MM-DD)
 
--   Go To Definition now works for class names too.
+-   _Go To Definition_ now works for class names too.
 
 ## 1.13.0 (2021-02-12)