+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## X.X.X (YYYY-MM-DD)
 
 -   _Go To Definition_ now works for class names too.
+-   _Go To Definition_ is now able to go in the `predefinedObject.xml` file set by the user with the option `yseopml.pathToPredefinedObjectsXml`.
 
 ## 1.13.0 (2021-02-12)