+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Release Notes
 
-## X.X.X (YYYY-MM-DD)
+## 1.14.0 (2021-06-04)
 
 -   _Go To Definition_ now works for class names too.
 -   _Go To Definition_ is now able to go in the `predefinedObject.xml` file set by the user with the option `yseopml.pathToPredefinedObjectsXml`.